+++ Assets/Scripts/_Data.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d730b4cf6a5384a85bbaaa2999003b9f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
