+++ Assets/Scripts/_Data/NoiseMap.cs
@@ -1,0 +1,10 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class NoiseMap
+{
+    public float[,] Map;
+    public float MinValue;
+    public float MaxValue;
+}
