+++ Assets/Scripts/_Data/Query.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6d7d58129948742a991af2c1b71e574e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
