+++ Assets/Scripts/_Data/Wave.cs
@@ -1,0 +1,11 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+[System.Serializable]
+public struct Wave
+{
+    public float Seed;
+    public float Frequency;
+    public float Amplitude;
+}
