+++ Assets/Scripts/_Data/Wave.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6c182216d2e4f4625b9b00f5b56ebe14
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
