+++ Assets/Scripts/_Data/WorldInstantiateMode.cs
@@ -1,0 +1,12 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+[System.Serializable]
+public enum WorldInstantiateMode
+{
+    None,
+    Height,
+    Moisture,
+    Heat,
+}
