+++ Assets/Scripts/_Data/WorldInstantiateMode.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: fe5067af979e64cf6b3230ea00b2b266
+guid: 221c53dd34936410f8421a15ef58948e
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
