+++ Assets/Scripts/_Data/WorldMap.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ec35c74f3e01646778c0f8cda4d05e70
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
